# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sentinel",
    "description": "Return Abuse Detection with Cost-Weighted Decisioning",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/internal/orders": {
      "get": {
        "tags": [
          "internal"
        ],
        "summary": "List Orders",
        "operationId": "list_orders_api_v1_internal_orders_get",
        "parameters": [
          {
            "name": "action",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Action"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Action"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "AUTO_APPLIED",
                    "PENDING_REVIEW",
                    "OVERRIDDEN",
                    "APPEAL_OPEN"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "DEMO",
                    "BACKTEST_REPLAY",
                    "LIVE"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "name": "min_p_abuse",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 1.0,
                  "minimum": 0.0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min P Abuse"
            }
          },
          {
            "name": "min_value_inr",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Value Inr"
            }
          },
          {
            "name": "graph_evidence",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "ANY",
                "STRONG",
                "WEAK_ONLY",
                "NONE"
              ],
              "type": "string",
              "default": "ANY",
              "title": "Graph Evidence"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "scored_at_desc",
                "p_abuse_desc",
                "value_desc",
                "exposure_desc"
              ],
              "type": "string",
              "default": "scored_at_desc",
              "title": "Sort"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "x-internal-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/orders/{order_id}": {
      "get": {
        "tags": [
          "internal"
        ],
        "summary": "Order Detail",
        "operationId": "order_detail_api_v1_internal_orders__order_id__get",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          },
          {
            "name": "x-internal-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/orders/{order_id}/override": {
      "post": {
        "tags": [
          "internal"
        ],
        "summary": "Override",
        "operationId": "override_api_v1_internal_orders__order_id__override_post",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          },
          {
            "name": "x-internal-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Key"
            }
          },
          {
            "name": "x-reviewer-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "reviewer-placeholder-01",
              "title": "X-Reviewer-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OverrideRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverrideResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/orders/{order_id}/appeal": {
      "post": {
        "tags": [
          "internal"
        ],
        "summary": "Appeal",
        "description": "Placeholder lifecycle (§14.2): opens an appeal and returns its APPEAL_OPENED audit event.",
        "operationId": "appeal_api_v1_internal_orders__order_id__appeal_post",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          },
          {
            "name": "x-internal-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Key"
            }
          },
          {
            "name": "x-reviewer-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "reviewer-placeholder-01",
              "title": "X-Reviewer-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppealRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditEventOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/audit-events": {
      "get": {
        "tags": [
          "internal"
        ],
        "summary": "List Audit Events",
        "operationId": "list_audit_events_api_v1_internal_audit_events_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "order_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order Id"
            }
          },
          {
            "name": "x-internal-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditEventsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/audit-events/verify": {
      "get": {
        "tags": [
          "internal"
        ],
        "summary": "Verify Audit Chain",
        "operationId": "verify_audit_chain_api_v1_internal_audit_events_verify_get",
        "parameters": [
          {
            "name": "x-internal-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditVerifyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/internal/demo/presets": {
      "get": {
        "tags": [
          "demo"
        ],
        "summary": "Presets",
        "description": "The three §11 demo requests, as seed-db wrote them beside the database.",
        "operationId": "presets_api_v1_internal_demo_presets_get",
        "parameters": [
          {
            "name": "x-internal-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScoreOrderRequest"
                  },
                  "title": "Response Presets Api V1 Internal Demo Presets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/demo/reset": {
      "post": {
        "tags": [
          "demo"
        ],
        "summary": "Reset",
        "operationId": "reset_api_v1_internal_demo_reset_post",
        "parameters": [
          {
            "name": "x-internal-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoResetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Action": {
        "type": "string",
        "enum": [
          "ALLOW",
          "PREPAID_ONLY",
          "MANUAL_REVIEW",
          "BLOCK"
        ],
        "title": "Action"
      },
...
      "AppealRequest": {
        "properties": {
          "channel": {
            "type": "string",
            "enum": [
              "CUSTOMER_SUPPORT",
              "EMAIL"
            ],
            "title": "Channel"
          },
          "note": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 10,
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "channel",
          "note"
        ],
        "title": "AppealRequest"
      },
      "AuditEventOut": {
        "properties": {
          "seq": {
            "type": "integer",
            "title": "Seq"
          },
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "event_type": {
            "type": "string",
            "enum": [
              "DECISION_CREATED",
              "OVERRIDE_APPLIED",
              "APPEAL_OPENED"
            ],
            "title": "Event Type"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "order_id": {
            "type": "string",
            "title": "Order Id"
          },
          "actor_type": {
            "type": "string",
            "enum": [
              "SYSTEM",
              "REVIEWER"
            ],
            "title": "Actor Type"
          },
          "actor_id": {
            "type": "string",
            "title": "Actor Id"
          },
          "previous_action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Action"
              },
              {
                "type": "null"
              }
            ]
          },
          "new_action": {
            "$ref": "#/components/schemas/Action"
          },
          "policy_version": {
            "type": "string",
            "title": "Policy Version"
          },
          "return_model_version": {
            "type": "string",
            "title": "Return Model Version"
          },
          "abuse_model_version": {
            "type": "string",
            "title": "Abuse Model Version"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "prev_hash": {
            "type": "string",
            "title": "Prev Hash"
          },
          "event_hash": {
            "type": "string",
            "title": "Event Hash"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "seq",
          "event_id",
          "event_type",
          "occurred_at",
          "order_id",
          "actor_type",
          "actor_id",
          "previous_action",
          "new_action",
          "policy_version",
          "return_model_version",
          "abuse_model_version",
          "payload",
          "prev_hash",
          "event_hash"
        ],
        "title": "AuditEventOut"
      },
      "AuditEventsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditEventOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "AuditEventsResponse"
      },
      "AuditVerifyResponse": {
        "properties": {
          "valid": {
            "type": "boolean",
            "title": "Valid"
          },
          "events_checked": {
            "type": "integer",
            "title": "Events Checked"
          },
          "first_broken_seq": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Broken Seq"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "valid",
          "events_checked",
          "first_broken_seq"
        ],
        "title": "AuditVerifyResponse"
      },
...
      "DemoResetResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "decisions": {
            "type": "integer",
            "title": "Decisions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "decisions"
        ],
        "title": "DemoResetResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrderDetailResponse": {
        "properties": {
          "order": {
            "$ref": "#/components/schemas/OrderSummary"
          },
          "customer": {
            "$ref": "#/components/schemas/CustomerSummary"
          },
          "decision": {
            "$ref": "#/components/schemas/ScoreOrderResponse"
          },
          "graph": {
            "$ref": "#/components/schemas/GraphPayload"
          },
          "baselines": {
            "items": {
              "$ref": "#/components/schemas/BaselineOutcome"
            },
            "type": "array",
            "title": "Baselines"
          },
          "audit_events": {
            "items": {
              "$ref": "#/components/schemas/AuditEventOut"
            },
            "type": "array",
            "title": "Audit Events"
          },
          "current_action": {
            "$ref": "#/components/schemas/Action"
          },
          "appeal_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appeal Reference"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "order",
          "customer",
          "decision",
          "graph",
          "baselines",
          "audit_events",
          "current_action",
          "appeal_reference"
        ],
        "title": "OrderDetailResponse"
      },
...
      "OverrideRequest": {
        "properties": {
          "new_action": {
            "$ref": "#/components/schemas/Action"
          },
          "reason_category": {
            "type": "string",
            "enum": [
              "CUSTOMER_VERIFIED",
              "INDEPENDENT_EVIDENCE_OF_ABUSE",
              "FALSE_POSITIVE_SHARED_IDENTIFIER",
              "POLICY_EXCEPTION",
              "OTHER"
            ],
            "title": "Reason Category"
          },
          "reason_text": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 15,
            "title": "Reason Text"
          },
          "expected_current_action": {
            "$ref": "#/components/schemas/Action"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "new_action",
          "reason_category",
          "reason_text",
          "expected_current_action"
        ],
        "title": "OverrideRequest"
      },
      "OverrideResponse": {
        "properties": {
          "order_id": {
            "type": "string",
            "title": "Order Id"
          },
          "decision_id": {
            "type": "string",
            "title": "Decision Id"
          },
          "original_recommendation": {
            "$ref": "#/components/schemas/Action"
          },
          "previous_action": {
            "$ref": "#/components/schemas/Action"
          },
          "new_action": {
            "$ref": "#/components/schemas/Action"
          },
          "reviewer_id": {
            "type": "string",
            "title": "Reviewer Id"
          },
          "overridden_at": {
            "type": "string",
            "format": "date-time",
            "title": "Overridden At"
          },
          "audit_event_id": {
            "type": "string",
            "title": "Audit Event Id"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "order_id",
          "decision_id",
          "original_recommendation",
          "previous_action",
          "new_action",
          "reviewer_id",
          "overridden_at",
          "audit_event_id",
          "warnings"
        ],
        "title": "OverrideResponse"
      },
...
      "QueueResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/QueueItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "counts_by_action": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/Action"
            },
            "type": "object",
            "title": "Counts By Action"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total",
          "counts_by_action"
        ],
        "title": "QueueResponse"
      },
...
      "ScoreOrderRequest": {
        "properties": {
          "order_id": {
            "type": "string",
            "pattern": "^ORD-[A-Z0-9-]{3,40}$",
            "title": "Order Id"
          },
          "account_id": {
            "type": "string",
            "pattern": "^ACC-[A-Z0-9-]{3,40}$",
            "title": "Account Id"
          },
          "placed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Placed At"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/OrderLineIn"
            },
            "type": "array",
            "maxItems": 50,
            "minItems": 1,
            "title": "Lines"
          },
          "discount_pct": {
            "type": "number",
            "maximum": 90.0,
            "minimum": 0.0,
            "title": "Discount Pct"
          },
          "delivery_speed": {
            "type": "string",
            "enum": [
              "STANDARD",
              "EXPRESS"
            ],
            "title": "Delivery Speed"
          },
          "payment_method": {
            "type": "string",
            "enum": [
              "PREPAID_CARD",
              "PREPAID_UPI",
              "COD"
            ],
            "title": "Payment Method"
          },
          "device_id": {
            "type": "string",
            "pattern": "^[a-f0-9]{32}$",
            "title": "Device Id"
          },
          "address_id": {
            "type": "string",
            "pattern": "^[a-f0-9]{32}$",
            "title": "Address Id"
          },
          "payment_token_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[a-f0-9]{32}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Token Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "order_id",
          "account_id",
          "placed_at",
          "lines",
          "discount_pct",
          "delivery_speed",
          "payment_method",
          "device_id",
          "address_id"
        ],
        "title": "ScoreOrderRequest",
        "description": "Internal order representation at t0 = after payment-method selection, before confirmation."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}